Fix hit flag bug.  Update macro
