+++ TrkDiag/test/CHCuts.C
@@ -1,0 +1,15 @@
+TCut pri("prel==0&&mcmom.R()>100");
+TCut prirel("prel>0");
+TCut Had("prel<0&&abs(mcpdg)>=20");
+TCut lowEe("prel<0&&abs(mcpdg)==11&&mcmom.R()<10.0");
+TCut other = TCut("prel<0")+(!Had)+(!lowEe);
+TCut multi("nch>1");
+TCut esel("esel");
+TCut tsel("tsel");
+TCut rsel("rsel");
+TCut nsel("nsel");
+TCut bkg("bkg==0");
+TCut bkgc("bkgclust");
+TCut sth("nch>1&&level==1");
+TCut ph("nsh>1&&level!=1");
+TCut sh("nch==1&&nsh==1");
