+++ TrkDiag/test/CHTest.C
@@ -1,0 +1,690 @@
+#include "TTree.h"
+#include "TCanvas.h"
+#include "TH1F.h"
+#include "TF1.h"
+#include "TCut.h"
+#include "TLegend.h"
+#include "THStack.h"
+#include <string>
+
+void ComboHitTest(TTree* CHD, const char* page="sel"){
+  string spage(page);
+  if(spage == "count" ) {
+    TH1F* nshc = new TH1F("nshc","N Straw Hits",8,0.5,8.5);
+    TH1F* nshh = new TH1F("nshh","N Straw Hits",8,0.5,8.5);
+    TH1F* nshb = new TH1F("nshb","N Straw Hits",8,0.5,8.5);
+    TH1F* nchc = new TH1F("nchc","N Combo Hits",8,0.5,8.5);
+    TH1F* nchh = new TH1F("nchh","N Combo Hits",8,0.5,8.5);
+    TH1F* nchb = new TH1F("nchb","N Combo Hits",8,0.5,8.5);
+    nshc->SetFillColorAlpha(kGreen,0.5);
+    nshb->SetFillColorAlpha(kRed,0.5);
+    nshh->SetFillColorAlpha(kBlue,0.5);
+    nchc->SetFillColorAlpha(kGreen,0.5);
+    nchb->SetFillColorAlpha(kRed,0.5);
+    nchh->SetFillColorAlpha(kBlue,0.5);
+    nshc->SetStats(0);
+    nshh->SetStats(0);
+    nshb->SetStats(0);
+    nchc->SetStats(0);
+    nchh->SetStats(0);
+    nchb->SetStats(0);
+    TCut goodmc("mcfrac==1.0");
+    CHD->Project("nshc","nsh",pri+goodmc);
+    CHD->Project("nshh","nsh",Had+goodmc);
+    CHD->Project("nshb","nsh",lowEe+goodmc);
+    CHD->Project("nchc","nch",pri+goodmc);
+    CHD->Project("nchh","nch",Had+goodmc);
+    CHD->Project("nchb","nch",lowEe+goodmc);
+
+    THStack* nsh = new THStack("nsh","N Straw Hits");
+    nsh->Add(nshc);
+    nsh->Add(nshh);
+    nsh->Add(nshb);
+    THStack* nch = new THStack("nch","N Combo Hits");
+    nch->Add(nchc);
+    nch->Add(nchh);
+    nch->Add(nchb);
+    TCanvas* ccan = new TCanvas("ccan","ccan",1000,500);
+    TLegend* cleg = new TLegend(0.5,0.7,0.8,0.9);
+    cleg->AddEntry(nshb,"e^{#pm} Bkg","F");
+    cleg->AddEntry(nshh,"Hadrons","F");
+    cleg->AddEntry(nshc,"Primary","F");
+    ccan->Divide(2,1);
+    ccan->cd(1);
+    gPad->SetLogy();
+    nsh->Draw();
+    cleg->Draw();
+    ccan->cd(2);
+    gPad->SetLogy();
+    nch->Draw();
+    cleg->Draw();
+
+  } else if(spage == "sep") {
+    TCut gs("nch>1&&_rel[0]==0&&_rel[1]==0");
+    TCut ms("nch>1&&(_rel[0]>0||_rel[1]>0)");
+    TCut bs("nch>1&&(_rel[0]<0 || _rel[1]<0)");
+
+    TH1F* gssep = new TH1F("gssep","Straw Id Separation;#Delta Straw Id",10,-0.5,9.5);
+    TH1F* mssep = new TH1F("mssep","Straw Id Separation;#Delta Straw Id",10,-0.5,9.5);
+    TH1F* bssep = new TH1F("bssep","Straw Id Separation;#Delta Straw Id",10,-0.5,9.5);
+    gssep->SetStats(0);
+    mssep->SetStats(0);
+    bssep->SetStats(0);
+    gssep->SetFillColorAlpha(kGreen,0.5);
+    mssep->SetFillColorAlpha(kBlue,0.5);
+    bssep->SetFillColorAlpha(kRed,0.5);
+    CHD->Project("gssep","_strawid[1]-_strawid[0]",gs);
+    CHD->Project("mssep","_strawid[1]-_strawid[0]",ms);
+    CHD->Project("bssep","_strawid[1]-_strawid[0]",bs);
+    THStack* ssep = new THStack("ssep","Straw Id Separation;#Delta Straw Id");
+    ssep->Add(bssep);
+    ssep->Add(mssep);
+    ssep->Add(gssep);
+    TH1F* gtsep = new TH1F("gtsep","Time Separation;#Delta t (ns)",70,0,70);
+    TH1F* mtsep = new TH1F("mtsep","Time Separation;#Delta t (ns)",70,0,70);
+    TH1F* btsep = new TH1F("btsep","Time Separation;#Delta t (ns)",70,0,70);
+    gtsep->SetStats(0);
+    mtsep->SetStats(0);
+    btsep->SetStats(0);
+    gtsep->SetFillColorAlpha(kGreen,0.5);
+    mtsep->SetFillColorAlpha(kBlue,0.5);
+    btsep->SetFillColorAlpha(kRed,0.5);
+    CHD->Project("gtsep","abs(_thit[1]-_thit[0])",gs);
+    CHD->Project("mtsep","abs(_thit[1]-_thit[0])",ms);
+    CHD->Project("btsep","abs(_thit[1]-_thit[0])",bs);
+    THStack* tsep = new THStack("tsep","Time Separation;#Delta t (ns)");
+    tsep->Add(btsep);
+    tsep->Add(mtsep);
+    tsep->Add(gtsep);
+
+    TH1F* gusep = new TH1F("gusep","#Delta U Pull;#Delta U/U_{err}",100,0,6);
+    TH1F* musep = new TH1F("musep","#Delta U Pull;#Delta U/U_{err}",100,0,6);
+    TH1F* busep = new TH1F("busep","#Delta U Pull;#Delta U/h_{err}",100,0,6);
+    gusep->SetStats(0);
+    musep->SetStats(0);
+    busep->SetStats(0);
+    gusep->SetFillColorAlpha(kGreen,0.5);
+    musep->SetFillColorAlpha(kBlue,0.5);
+    busep->SetFillColorAlpha(kRed,0.5);
+    CHD->Project("gusep","_du/_ures",gs);
+    CHD->Project("musep","_du/_ures",ms);
+    CHD->Project("busep","_du/_ures",bs);
+    THStack* usep = new THStack("usep","#Delta U Pull;#Delta U/h_{err}");
+    usep->Add(busep);
+    usep->Add(musep);
+    usep->Add(gusep);
+
+    TCanvas* sepcan = new TCanvas("sepcan","sepcan",1000,1000);
+    sepcan->Divide(2,2);
+    sepcan->cd(1);
+    gPad->SetLogy();
+    ssep->Draw();
+    sepcan->cd(2);
+    gPad->SetLogy();
+    tsep->Draw();
+    sepcan->cd(3);
+    gPad->SetLogy();
+    usep->Draw();
+    sepcan->cd(4);
+    TLegend* sepleg = new TLegend(0.3,0.3,0.8,0.8);
+    sepleg->AddEntry(gssep,"Same Particle","F");
+    sepleg->AddEntry(mssep,"Related Particle","F");
+    sepleg->AddEntry(bssep,"Unrelated Particle","F");
+    sepleg->Draw();
+
+  } else if(spage == "qual") {
+
+    TH1F* gqualc = new TH1F("gqualc","Combo Quality",100,0,6.0);
+    TH1F* gqualh = new TH1F("gqualh","Combo Quality",100,0,6.0);
+    TH1F* gqualb = new TH1F("gqualb","Combo Quality",100,0,6.0);
+    TH1F* bqualc = new TH1F("bqualc","Combo Quality",100,0,6.0);
+    TH1F* bqualh = new TH1F("bqualh","Combo Quality",100,0,6.0);
+    TH1F* bqualb = new TH1F("bqualb","Combo Quality",100,0,6.0);
+    gqualc->SetFillColorAlpha(kGreen,0.5);
+    gqualb->SetFillColorAlpha(kGreen,0.5);
+    gqualh->SetFillColorAlpha(kGreen,0.5);
+    bqualc->SetFillColorAlpha(kRed,0.5);
+    bqualb->SetFillColorAlpha(kRed,0.5);
+    bqualh->SetFillColorAlpha(kRed,0.5);
+    gqualc->SetStats(0);
+    gqualh->SetStats(0);
+    gqualb->SetStats(0);
+    bqualc->SetStats(0);
+    bqualh->SetStats(0);
+    bqualb->SetStats(0);
+
+    TCut gc("mcfrac==1.0");
+    TCut bc("mcfrac<1.0");
+    CHD->Project("gqualc","qual",pri+gc);
+    CHD->Project("gqualh","qual",Had+gc);
+    CHD->Project("gqualb","qual",lowEe+gc);
+    CHD->Project("bqualc","qual",pri+bc);
+    CHD->Project("bqualh","qual",Had+bc);
+    CHD->Project("bqualb","qual",lowEe+bc);
+    THStack* qualc = new THStack("qualc","Combo Quality, Primary");
+    qualc->Add(bqualc);
+    qualc->Add(gqualc);
+    gqualc->SetMinimum(0.1);
+    bqualc->SetMinimum(0.1);
+    THStack* qualh = new THStack("qualh","Combo Quality, Hadron");
+    qualh->Add(bqualh);
+    qualh->Add(gqualh);
+    THStack* qualb = new THStack("qualb","Combo Quality, Low-E e^{#pm}");
+    qualb->Add(bqualb);
+    qualb->Add(gqualb);
+
+    TCanvas* qcan = new TCanvas("qcan","qcan",1000,1000);
+    qcan->Divide(2,2);
+    qcan->cd(1);
+    gPad->SetLogy();
+    qualc->Draw();
+    qcan->cd(2);
+    gPad->SetLogy();
+    qualh->Draw();
+    qcan->cd(3);
+    gPad->SetLogy();
+    qualb->Draw();
+
+    TLegend* qleg = new TLegend(0.5,0.6,0.9,0.9);
+    qleg->AddEntry(gqualc,"Single-particle Combo","F");
+    qleg->AddEntry(bqualc,"Mixed Combo","F");
+    qleg->Draw();
+
+  } else if(spage == "sel" ) {
+    TH1F* hctimec = new TH1F("hctimec","Combo Hit Corrected Time;Hit Time (ns)",500,350,1750);
+    TH1F* hctimeb = new TH1F("hctimeb","Combo Hit Corrected Time;Hit Time (ns)",500,350,1750);
+    TH1F* hctimeh = new TH1F("hctimeh","Combo Hit Corrected Time;Hit Time (ns)",500,350,1750);
+    TH1F* sctimec = new TH1F("sctimec","Combo Hit Corrected Time;Hit Time (ns)",500,350,1750);
+    TH1F* sctimeb = new TH1F("sctimeb","Combo Hit Corrected Time;Hit Time (ns)",500,350,1750);
+    TH1F* sctimeh = new TH1F("sctimeh","Combo Hit Corrected Time;Hit Time (ns)",500,350,1750);
+    hctimec->SetLineColor(kGreen);
+    hctimeb->SetLineColor(kRed);
+    hctimeh->SetLineColor(kBlue);
+    sctimec->SetFillColorAlpha(kGreen,0.1);
+    sctimeb->SetFillColorAlpha(kRed,0.1);
+    sctimeh->SetFillColorAlpha(kBlue,0.1);
+    hctimec->SetStats(0);
+    hctimeh->SetStats(0);
+    hctimeb->SetStats(0);
+    sctimec->SetStats(0);
+    sctimeh->SetStats(0);
+    sctimeb->SetStats(0);
+    CHD->Project("hctimec","ctime","nsh"*pri);
+    CHD->Project("hctimeb","ctime","nsh"*lowEe);
+    CHD->Project("hctimeh","ctime","nsh"*Had);
+    CHD->Project("sctimec","ctime","nsh"*(pri+tsel));
+    CHD->Project("sctimeb","ctime","nsh"*(lowEe+tsel));
+    CHD->Project("sctimeh","ctime","nsh"*(Had+tsel));
+    TH1F* hrhoc = new TH1F("hrhoc","Combo Hit Transverse Radius;Hit #rho (mm)",100,350.0,800.0);
+    TH1F* hrhob = new TH1F("hrhob","Combo Hit Transverse Radius;Hit #rho (mm)",100,350.0,800.0);
+    TH1F* hrhoh = new TH1F("hrhoh","Combo Hit Transverse Radius;Hit #rho (mm)",100,350.0,800.0);
+    TH1F* srhoc = new TH1F("srhoc","Combo Hit Transverse Radius;Hit #rho (mm)",100,350.0,800.0);
+    TH1F* srhob = new TH1F("srhob","Combo Hit Transverse Radius;Hit #rho (mm)",100,350.0,800.0);
+    TH1F* srhoh = new TH1F("srhoh","Combo Hit Transverse Radius;Hit #rho (mm)",100,350.0,800.0);
+    hrhoc->SetLineColor(kGreen);
+    hrhob->SetLineColor(kRed);
+    hrhoh->SetLineColor(kBlue);
+    srhoc->SetFillColorAlpha(kGreen,0.1);
+    srhob->SetFillColorAlpha(kRed,0.1);
+    srhoh->SetFillColorAlpha(kBlue,0.1);
+    hrhoc->SetStats(0);
+    hrhoh->SetStats(0);
+    hrhob->SetStats(0);
+    srhoc->SetStats(0);
+    srhoh->SetStats(0);
+    srhob->SetStats(0);
+    CHD->Project("hrhoc","pos.Rho()","nsh"*pri);
+    CHD->Project("hrhob","pos.Rho()","nsh"*lowEe);
+    CHD->Project("hrhoh","pos.Rho()","nsh"*Had);
+    CHD->Project("srhoc","pos.Rho()","nsh"*(pri+rsel));
+    CHD->Project("srhob","pos.Rho()","nsh"*(lowEe+rsel));
+    CHD->Project("srhoh","pos.Rho()","nsh"*(Had+rsel));
+    TH1F* hedepc = new TH1F("hedepc","Combo Hit EDep;Hit EDep (KeV)",100,-0.1,6.0);
+    TH1F* hedepb = new TH1F("hedepb","Combo Hit EDep;Hit EDep (KeV)",100,-0.1,6.0);
+    TH1F* hedeph = new TH1F("hedeph","Combo Hit EDep;Hit EDep (KeV)",100,-0.1,6.0);
+    TH1F* sedepc = new TH1F("sedepc","Combo Hit EDep;Hit EDep (KeV)",100,-0.1,6.0);
+    TH1F* sedepb = new TH1F("sedepb","Combo Hit EDep;Hit EDep (KeV)",100,-0.1,6.0);
+    TH1F* sedeph = new TH1F("sedeph","Combo Hit EDep;Hit EDep (KeV)",100,-0.1,6.0);
+    hedepc->SetLineColor(kGreen);
+    hedepb->SetLineColor(kRed);
+    hedeph->SetLineColor(kBlue);
+    sedepc->SetFillColorAlpha(kGreen,0.1);
+    sedepb->SetFillColorAlpha(kRed,0.1);
+    sedeph->SetFillColorAlpha(kBlue,0.1);
+    hedepc->SetStats(0);
+    hedepb->SetStats(0);
+    hedeph->SetStats(0);
+    sedepc->SetStats(0);
+    sedepb->SetStats(0);
+    sedeph->SetStats(0);
+    CHD->Project("hedepc","1000.0*edep","nsh"*pri);
+    CHD->Project("hedepb","1000.0*edep","nsh"*lowEe);
+    CHD->Project("hedeph","1000.0*edep","nsh"*Had);
+    CHD->Project("sedepc","1000.0*edep","nsh"*(pri+esel));
+    CHD->Project("sedepb","1000.0*edep","nsh"*(lowEe+esel));
+    CHD->Project("sedeph","1000.0*edep","nsh"*(Had+esel));
+
+    TH1F* all = new TH1F("all","Net Hit Selection Efficiency",3,0.5,3.5);
+    TH1F* sel = new TH1F("sel","Net Hit Selection Efficiency",3,0.5,3.5);
+    TH1F* eff = new TH1F("eff","Net Hit Selection Efficiency",3,0.5,3.5);
+    eff->SetStats(0);
+    eff->SetLabelSize(0.1);
+    eff->GetXaxis()->SetBinLabel(1,"Primary");
+    eff->GetXaxis()->SetBinLabel(2,"Low-E e^{#pm}");
+    eff->GetXaxis()->SetBinLabel(3,"Hadrons");
+
+
+    CHD->Project("sel","1.0","nsh"*(pri+esel+tsel+rsel));
+    CHD->Project("+sel","2.0","nsh"*(lowEe+esel+tsel+rsel));
+    CHD->Project("+sel","3.0","nsh"*(Had+esel+tsel+rsel));
+    CHD->Project("all","1.0","mcndigi"*(pri));
+    CHD->Project("+all","2.0","mcndigi"*(lowEe));
+    CHD->Project("+all","3.0","mcndigi"*(Had));
+
+    eff->Divide(sel,all);
+    TPaveText *seltxt = new TPaveText(0.15,0.15,0.6,0.5,"NDC");
+    char line[80];
+    snprintf(line,80,"Primary Efficiency = %3.3f",eff->GetBinContent(1));
+    seltxt->AddText(line);
+    snprintf(line,80,"Low-E e^{#pm} Efficiency = %3.3f",eff->GetBinContent(2));
+    seltxt->AddText(line);
+    snprintf(line,80,"Hadron Efficiency = %3.3f",eff->GetBinContent(3));
+    seltxt->AddText(line);
+    snprintf(line,80,"Primary Purity = %3.3f",sel->GetBinContent(1)/(sel->GetBinContent(1)+sel->GetBinContent(2)+sel->GetBinContent(3)));
+    seltxt->AddText(line);
+
+//    double beff,heff,ceff;
+//    char legtxt[50];
+//    TLegend* ctimeleg = new TLegend(0.2,0.1,0.6,0.3);
+//    beff = sctimeb->GetEntries()/hctimeb->GetEntries();
+//    heff = sctimeh->GetEntries()/hctimeh->GetEntries();
+//    ceff = sctimec->GetEntries()/hctimec->GetEntries();
+//    snprintf(legtxt,50,"e^{#pm} Bkg, #epsilon = %3.3f",beff);
+//    ctimeleg->AddEntry(hctimeb,legtxt,"L");
+//    snprintf(legtxt,50,"Hadrons, #epsilon = %3.3f",heff);
+//    ctimeleg->AddEntry(hctimeh,legtxt,"L");
+//    snprintf(legtxt,50,"pri, #epsilon = %3.3f",ceff);
+//    ctimeleg->AddEntry(hctimec,legtxt,"L");
+//
+//    TLegend* rholeg = new TLegend(0.2,0.1,0.6,0.3);
+//    beff = srhob->GetEntries()/hrhob->GetEntries();
+//    heff = srhoh->GetEntries()/hrhoh->GetEntries();
+//    ceff = srhoc->GetEntries()/hrhoc->GetEntries();
+//    snprintf(legtxt,50,"e^{#pm} Bkg, #epsilon = %3.3f",beff);
+//    rholeg->AddEntry(hrhob,legtxt,"L");
+//    snprintf(legtxt,50,"Hadrons, #epsilon = %3.3f",heff);
+//    rholeg->AddEntry(hrhoh,legtxt,"L");
+//    snprintf(legtxt,50,"pri, #epsilon = %3.3f",ceff);
+//    rholeg->AddEntry(hrhoc,legtxt,"L");
+//
+//    TLegend* edepleg = new TLegend(0.2,0.1,0.6,0.3);
+//    beff = sedepb->GetEntries()/hedepb->GetEntries();
+//    heff = sedeph->GetEntries()/hedeph->GetEntries();
+//    ceff = sedepc->GetEntries()/hedepc->GetEntries();
+//    snprintf(legtxt,50,"e^{#pm} Bkg, #epsilon = %3.3f",beff);
+//    edepleg->AddEntry(hedepb,legtxt,"L");
+//    snprintf(legtxt,50,"Hadrons, #epsilon = %3.3f",heff);
+//    edepleg->AddEntry(hedeph,legtxt,"L");
+//    snprintf(legtxt,50,"pri, #epsilon = %3.3f",ceff);
+//    edepleg->AddEntry(hedepc,legtxt,"L");
+//
+    TLegend* selleg = new TLegend(0.3,0.1,0.6,0.4);
+    selleg->AddEntry(hedepb,"Low-E e^{#pm}","LF");
+    selleg->AddEntry(hedeph,"Hadrons","LF");
+    selleg->AddEntry(hedepc,"Primary","LF");
+
+    TCanvas* scan = new TCanvas("scan","scan",1500,1000);
+    scan->Divide(2,2);
+    scan->cd(1);
+    gPad->SetLogy();
+    hctimeb->Draw();
+    hctimeb->SetMinimum(1);
+    hctimeh->Draw("same");
+    hctimec->Draw("same");
+    sctimeb->Draw("sameh");
+    sctimeh->Draw("sameh");
+    sctimec->Draw("sameh");
+ //   ctimeleg->Draw();
+    scan->cd(2);
+    gPad->SetLogy();
+    hrhob->SetMinimum(1);
+    hrhob->Draw();
+    hrhoh->Draw("same");
+    hrhoc->Draw("same");
+    srhob->Draw("sameh");
+    srhoh->Draw("sameh");
+    srhoc->Draw("sameh");
+ //   rholeg->Draw();
+    scan->cd(3);
+    gPad->SetLogy();
+    hedepb->SetMinimum(1);
+    hedepb->Draw();
+    hedeph->Draw("same");
+    hedepc->Draw("same");
+    sedeph->Draw("sameh");
+    sedepb->Draw("sameh");
+    sedepc->Draw("sameh");
+    selleg->Draw();
+    scan->cd(4);
+    eff->Draw("h");
+    seltxt->Draw();
+
+  } else if(spage =="pres") {
+    TH1F* rresshc = new TH1F("rresshc","Straw Hit Radial Resolution;#rho_{reco}-#rho_{MC} (mm)",100,-50,50);
+    TH1F* rresshb = new TH1F("rresshb","Straw Hit Radial Resolution;#rho_{reco}-#rho_{MC} (mm)",100,-50,50);
+    TH1F* rresshh = new TH1F("rresshh","Straw Hit Radial Resolution;#rho_{reco}-#rho_{MC} (mm)",100,-50,50);
+    TH1F* fresshc = new TH1F("fresshc","Straw Hit Azimuthal Resolution;#rho#times(#phi_{reco}-#phi_{MC}) (mm)",100,-200,200);
+    TH1F* fresshb = new TH1F("fresshb","Straw Hit Azimuthal Resolution;#rho#times(#phi_{reco}-#phi_{MC}) (mm)",100,-200,200);
+    TH1F* fresshh = new TH1F("fresshh","Straw Hit Azimuthal Resolution;#rho#times(#phi_{reco}-#phi_{MC}) (mm)",100,-200,200);
+    rresshc->SetLineColor(kGreen);
+    rresshb->SetLineColor(kRed);
+    rresshh->SetLineColor(kBlue);
+    fresshc->SetLineColor(kGreen);
+    fresshb->SetLineColor(kRed);
+    fresshh->SetLineColor(kBlue);
+    rresshc->SetStats(0);
+    rresshb->SetStats(0);
+    rresshh->SetStats(0);
+    fresshc->SetStats(0);
+    fresshb->SetStats(0);
+    fresshh->SetStats(0);
+
+    TH1F* rresphc = new TH1F("rresphc","Panel Hit Radial Resolution;#rho_{reco}-#rho_{MC} (mm)",100,-50,50);
+    TH1F* rresphb = new TH1F("rresphb","Panel Hit Radial Resolution;#rho_{reco}-#rho_{MC} (mm)",100,-50,50);
+    TH1F* rresphh = new TH1F("rresphh","Panel Hit Radial Resolution;#rho_{reco}-#rho_{MC} (mm)",100,-50,50);
+    TH1F* fresphc = new TH1F("fresphc","Panel Hit Azimuthal Resolution;#rho#times(#phi_{reco}-#phi_{MC}) (mm)",100,-200,200);
+    TH1F* fresphb = new TH1F("fresphb","Panel Hit Azimuthal Resolution;#rho#times(#phi_{reco}-#phi_{MC}) (mm)",100,-200,200);
+    TH1F* fresphh = new TH1F("fresphh","Panel Hit Azimuthal Resolution;#rho#times(#phi_{reco}-#phi_{MC}) (mm)",100,-200,200);
+    rresphc->SetLineColor(kGreen);
+    rresphb->SetLineColor(kRed);
+    rresphh->SetLineColor(kBlue);
+    fresphc->SetLineColor(kGreen);
+    fresphb->SetLineColor(kRed);
+    fresphh->SetLineColor(kBlue);
+    rresphc->SetStats(0);
+    rresphb->SetStats(0);
+    rresphh->SetStats(0);
+    fresphc->SetStats(0);
+    fresphb->SetStats(0);
+    fresphh->SetStats(0);
+
+    TH1F* rresstc = new TH1F("rresstc","Stereo Hit Radial Resolution;#rho_{reco}-#rho_{MC} (mm)",200,-50,50);
+    TH1F* rresstb = new TH1F("rresstb","Stereo Hit Radial Resolution;#rho_{reco}-#rho_{MC} (mm)",200,-50,50);
+    TH1F* rressth = new TH1F("rressth","Stereo Hit Radial Resolution;#rho_{reco}-#rho_{MC} (mm)",200,-50,50);
+    TH1F* fresstc = new TH1F("fresstc","Stereo Hit Azimuthal Resolution;#rho#times(#phi_{reco}-#phi_{MC}) (mm)",200,-200,200);
+    TH1F* fresstb = new TH1F("fresstb","Stereo Hit Azimuthal Resolution;#rho#times(#phi_{reco}-#phi_{MC}) (mm)",200,-200,200);
+    TH1F* fressth = new TH1F("fressth","Stereo Hit Azimuthal Resolution;#rho#times(#phi_{reco}-#phi_{MC}) (mm)",200,-200,200);
+    rresstc->SetLineColor(kGreen);
+    rresstb->SetLineColor(kRed);
+    rressth->SetLineColor(kBlue);
+    fresstc->SetLineColor(kGreen);
+    fresstb->SetLineColor(kRed);
+    fressth->SetLineColor(kBlue);
+    rresstc->SetStats(0);
+    rresstb->SetStats(0);
+    rressth->SetStats(0);
+    fresstc->SetStats(0);
+    fresstb->SetStats(0);
+    fressth->SetStats(0);
+
+    CHD->Project("rresshc","pos.Rho()-mcpos.Rho()","nsh"*(pri+sh));
+    CHD->Project("rresshb","pos.Rho()-mcpos.Rho()","nsh"*(lowEe+sh));
+    CHD->Project("rresshh","pos.Rho()-mcpos.Rho()","nsh"*(Had+sh));
+    rresshc->Scale(rresshb->Integral()/rresshc->Integral());
+    rresshh->Scale(rresshb->Integral()/rresshh->Integral());
+    CHD->Project("fresshc","mcpos.Rho()*(pos.Phi()-mcpos.Phi())","nsh"*(pri+sh));
+    CHD->Project("fresshb","mcpos.Rho()*(pos.Phi()-mcpos.Phi())","nsh"*(lowEe+sh));
+    CHD->Project("fresshh","mcpos.Rho()*(pos.Phi()-mcpos.Phi())","nsh"*(Had+sh));
+    fresshc->Scale(fresshb->Integral()/fresshc->Integral());
+    fresshh->Scale(fresshb->Integral()/fresshh->Integral());
+
+    CHD->Project("rresphc","pos.Rho()-mcpos.Rho()","nsh"*(pri+ph));
+    CHD->Project("rresphb","pos.Rho()-mcpos.Rho()","nsh"*(lowEe+ph));
+    CHD->Project("rresphh","pos.Rho()-mcpos.Rho()","nsh"*(Had+ph));
+    rresphc->Scale(rresphb->Integral()/rresphc->Integral());
+    rresphh->Scale(rresphb->Integral()/rresphh->Integral());
+    CHD->Project("fresphc","mcpos.Rho()*(pos.Phi()-mcpos.Phi())","nsh"*(pri+ph));
+    CHD->Project("fresphb","mcpos.Rho()*(pos.Phi()-mcpos.Phi())","nsh"*(lowEe+ph));
+    CHD->Project("fresphh","mcpos.Rho()*(pos.Phi()-mcpos.Phi())","nsh"*(Had+ph));
+    fresphc->Scale(fresphb->Integral()/fresphc->Integral());
+    fresphh->Scale(fresphb->Integral()/fresphh->Integral());
+
+    CHD->Project("rresstc","pos.Rho()-mcpos.Rho()","nsh"*(pri+sth));
+    CHD->Project("rresstb","pos.Rho()-mcpos.Rho()","nsh"*(lowEe+sth));
+    CHD->Project("rressth","pos.Rho()-mcpos.Rho()","nsh"*(Had+sth));
+    rresstc->Scale(rresstb->Integral()/rresstc->Integral());
+    rressth->Scale(rresstb->Integral()/rressth->Integral());
+    CHD->Project("fresstc","mcpos.Rho()*(pos.Phi()-mcpos.Phi())","nsh"*(pri+sth));
+    CHD->Project("fresstb","mcpos.Rho()*(pos.Phi()-mcpos.Phi())","nsh"*(lowEe+sth));
+    CHD->Project("fressth","mcpos.Rho()*(pos.Phi()-mcpos.Phi())","nsh"*(Had+sth));
+    fresstc->Scale(fresstb->Integral()/fresstc->Integral());
+    fressth->Scale(fresstb->Integral()/fressth->Integral());
+
+    TLegend* rleg = new TLegend(0.1,0.7,0.4,0.9);
+    rleg->AddEntry(rresshb,"e^{#pm} Bkg","L");
+    rleg->AddEntry(rresshh,"Hadrons (scaled)","L");
+    rleg->AddEntry(rresshc,"pri (scaled)","L");
+
+    TCanvas* pcan = new TCanvas("pcan","pcan",1200,1200);
+    pcan->Divide(2,3);
+    pcan->cd(1);
+    rresshh->Draw("h");
+    rresshc->Draw("sameh");
+    rresshb->Draw("sameh");
+    rleg->Draw();
+    pcan->cd(2);
+    fresshh->Draw("h");
+    fresshc->Draw("sameh");
+    fresshb->Draw("sameh");
+    pcan->cd(3);
+    rresphh->Draw("h");
+    rresphc->Draw("sameh");
+    rresphb->Draw("sameh");
+    rleg->Draw();
+    pcan->cd(4);
+    fresphh->Draw("h");
+    fresphc->Draw("sameh");
+    fresphb->Draw("sameh");
+    pcan->cd(5);
+    rresstb->Draw("h");
+    rresstc->Draw("sameh");
+    rressth->Draw("sameh");
+    rleg->Draw();
+    pcan->cd(6);
+    fresstb->Draw("h");
+    fresstc->Draw("sameh");
+    fressth->Draw("sameh");
+
+  } else if(spage =="ures") {
+    TH1F* uresc = new TH1F("uresc","U Pos Res;U_{reco}-U_{MC} (mm)",100,-300,300);
+    TH1F* uresb = new TH1F("uresb","U Pos Res;U_{reco}-U_{MC} (mm)",100,-300,300);
+    TH1F* uresh = new TH1F("uresh","U Pos Res;U_{reco}-U_{MC} (mm)",100,-300,300);
+    uresc->SetLineColor(kGreen);
+    uresb->SetLineColor(kRed);
+    uresh->SetLineColor(kBlue);
+    CHD->Project("uresc","udist-mcudist",pri);
+    CHD->Project("uresb","udist-mcudist",lowEe);
+    CHD->Project("uresh","udist-mcudist",Had);
+    uresc->Scale(uresb->GetEntries()/uresc->GetEntries());
+    uresh->Scale(uresb->GetEntries()/uresh->GetEntries());
+    TLegend* wleg = new TLegend(0.1,0.7,0.4,0.9);
+    wleg->AddEntry(uresb,"e^{#pm} Bkg","L");
+    wleg->AddEntry(uresh,"Hadrons","L");
+    wleg->AddEntry(uresc,"Primary","L");
+    TH1F* upullc = new TH1F("upullc","U Pos Pull;U_{reco}-U_{MC}/U_{err}",100,-10,10);
+    TH1F* upullb = new TH1F("upullb","U Pos Pull;U_{reco}-U_{MC}/U_{err}",100,-10,10);
+    TH1F* upullh = new TH1F("upullh","U Pos Pull;U_{reco}-U_{MC}/h_{err}",100,-10,10);
+
+    upullc->SetLineColor(kGreen);
+    upullb->SetLineColor(kRed);
+    upullh->SetLineColor(kBlue);
+    CHD->Project("upullc","(udist-mcudist)/ures",pri);
+    CHD->Project("upullb","(udist-mcudist)/ures",lowEe);
+    CHD->Project("upullh","(udist-mcudist)/ures",Had);
+    upullc->Scale(upullb->GetEntries()/upullc->GetEntries());
+    upullh->Scale(upullb->GetEntries()/upullh->GetEntries());
+    TCanvas* ucan = new TCanvas("ucan","ucan",1000,500);
+    ucan->Divide(2,1);
+    ucan->cd(1);
+    uresc->Fit("gaus","","",-80,80);
+    uresb->Draw("same");
+    uresh->Draw("same");
+    wleg->Draw();
+    ucan->cd(2);
+    upullc->Fit("gaus","","",-80,80);
+    upullb->Draw("samehs");
+    upullh->Draw("samehs");
+  } else if(spage =="timeres"){
+    TH1F* timeresc = new TH1F("timeresc","Time Res;t_{reco}-t_{MC} (nsec)",100,-50,50);
+    TH1F* timeresb = new TH1F("timeresb","Time Res;t_{reco}-t_{MC} (nsec)",100,-50,50);
+    TH1F* timeresh = new TH1F("timeresh","Time Res;t_{reco}-t_{MC} (nsec)",100,-50,50);
+    timeresc->SetLineColor(kGreen);
+    timeresb->SetLineColor(kRed);
+    timeresh->SetLineColor(kBlue);
+    CHD->Project("timeresc","ctime-fmod(mctime,1695)",pri);
+    CHD->Project("timeresb","ctime-fmod(mctime,1695)",lowEe);
+    CHD->Project("timeresh","ctime-fmod(mctime,1695)",Had);
+    timeresc->Scale(timeresb->GetEntries()/timeresc->GetEntries());
+    timeresh->Scale(timeresb->GetEntries()/timeresh->GetEntries());
+    TLegend* wleg = new TLegend(0.1,0.7,0.4,0.9);
+    wleg->AddEntry(timeresb,"e^{#pm} Bkg","L");
+    wleg->AddEntry(timeresh,"Hadrons","L");
+    wleg->AddEntry(timeresc,"Primary","L");
+    TH1F* timepullc = new TH1F("timepullc","Time Pull;t_{reco}-t_{MC}/t_{err}",100,-10,10);
+    TH1F* timepullb = new TH1F("timepullb","Time Pull;t_{reco}-t_{MC}/t_{err}",100,-10,10);
+    TH1F* timepullh = new TH1F("timepullh","Time Pull;t_{reco}-t_{MC}/t_{err}",100,-10,10);
+    timepullc->SetLineColor(kGreen);
+    timepullb->SetLineColor(kRed);
+    timepullh->SetLineColor(kBlue);
+    CHD->Project("timepullc","(ctime-fmod(mctime,1695))/tres",pri);
+    CHD->Project("timepullb","(ctime-fmod(mctime,1695))/tres",lowEe);
+    CHD->Project("timepullh","(ctime-fmod(mctime,1695))/tres",Had);
+    timepullc->Scale(timepullb->GetEntries()/timepullc->GetEntries());
+    timepullh->Scale(timepullb->GetEntries()/timepullh->GetEntries());
+    TCanvas* timecan = new TCanvas("timecan","timecan",1000,500);
+    timecan->Divide(2,1);
+    timecan->cd(1);
+    timeresc->Fit("gaus","","",-80,80);
+    timeresb->Draw("same");
+    timeresh->Draw("same");
+    wleg->Draw();
+    timecan->cd(2);
+    timepullc->Fit("gaus","","",-80,80);
+    timepullb->Draw("samehs");
+    timepullh->Draw("samehs");
+    wleg->Draw();
+  } else if(spage == "mcmatch") {
+    TH1F* mcpurc = new TH1F("mcpurc","ComboHit True Purity;N_{same MC}/N_{Combo}",20,0.0,1.01);
+    TH1F* mcpurb = new TH1F("mcpurb","ComboHit True Purity;N_{same MC}/N_{Combo}",20,0.0,1.01);
+    TH1F* mcpurh = new TH1F("mcpurh","ComboHit True Purity;N_{same MC}/N_{Combo}",20,0.0,1.01);
+    mcpurc->SetLineColor(kGreen);
+    mcpurb->SetLineColor(kRed);
+    mcpurh->SetLineColor(kBlue);
+    mcpurc->SetStats(0);
+    mcpurb->SetStats(0);
+    mcpurh->SetStats(0);
+    CHD->Project("mcpurc","mcfrac",pri);
+    CHD->Project("mcpurb","mcfrac",lowEe);
+    CHD->Project("mcpurh","mcfrac",Had);
+    mcpurc->Scale(1.0/mcpurc->GetEntries());
+    mcpurb->Scale(1.0/mcpurb->GetEntries());
+    mcpurh->Scale(1.0/mcpurh->GetEntries());
+    TH1F* mceffc = new TH1F("mceffc","ComboHit True Efficiency;N_{Combo}/N_{MC True In Panel}",20,0.0,1.01);
+    TH1F* mceffb = new TH1F("mceffb","ComboHit True Efficiency;N_{Combo}/N_{MC True In Panel}",20,0.0,1.01);
+    TH1F* mceffh = new TH1F("mceffh","ComboHit True Efficiency;N_{Combo}/N_{MC True In Panel}",20,0.0,1.01);
+    mceffc->SetLineColor(kGreen);
+    mceffb->SetLineColor(kRed);
+    mceffh->SetLineColor(kBlue);
+    mceffc->SetStats(0);
+    mceffb->SetStats(0);
+    mceffh->SetStats(0);
+    CHD->Project("mceffc","nsh/mcndigi",pri);
+    CHD->Project("mceffb","nsh/mcndigi",lowEe);
+    CHD->Project("mceffh","nsh/mcndigi",Had);
+    mceffc->Scale(1.0/mceffc->GetEntries());
+    mceffb->Scale(1.0/mceffb->GetEntries());
+    mceffh->Scale(1.0/mceffh->GetEntries());
+    TLegend* purleg = new TLegend(0.1,0.5,0.6,0.8);
+    char line[80];
+    snprintf(line,80,"e^{#pm} Bkg, <purity> = %3.3f",mcpurb->GetMean());
+    purleg->AddEntry(mcpurb,line,"L");
+    snprintf(line,80,"Hadron, <purity> = %3.3f",mcpurh->GetMean());
+    purleg->AddEntry(mcpurh,line,"L");
+    snprintf(line,80,"Primary, <purity> = %3.3f",mcpurc->GetMean());
+    purleg->AddEntry(mcpurc,line,"L");
+    TLegend* effleg = new TLegend(0.1,0.5,0.6,0.8);
+    snprintf(line,80,"e^{#pm} Bkg, <efficiency> = %3.3f",mceffb->GetMean());
+    effleg->AddEntry(mceffb,line,"L");
+    snprintf(line,80,"Hadron, <efficiency> = %3.3f",mceffh->GetMean());
+    effleg->AddEntry(mceffh,line,"L");
+    snprintf(line,80,"Primary, <efficiency> = %3.3f",mceffc->GetMean());
+    effleg->AddEntry(mceffc,line,"L");
+    TCanvas* mcmcan = new TCanvas("mcmcan","mcmcan",1200,600);
+    mcmcan->Divide(2,1);
+    mcmcan->cd(1);
+    mcpurc->Draw("h");
+    mcpurb->Draw("sameh");
+    mcpurh->Draw("sameh");
+    purleg->Draw();
+    mcmcan->cd(2);
+    mceffc->Draw("h");
+    mceffb->Draw("sameh");
+    mceffh->Draw("sameh");
+    effleg->Draw();
+
+  } else if(spage == "flag") {
+    static const size_t nsel(6);
+    static const size_t npart(5);
+    TH2F* norm = new TH2F("norm","Hit Selection",npart,-0.5,npart-0.5,nsel,-0.5,nsel-0.5);
+    TH2F* sel = new TH2F("sel","Hit Selection",npart,-0.5,npart-0.5,nsel,-0.5,nsel-0.5);
+    TH2F* selfrac = new TH2F("selfrac","Hit Selection Fraction",npart,-0.5,npart-0.5,nsel,-0.5,nsel-0.5);
+    norm->SetStats(0);
+    sel->SetStats(0);
+    selfrac->SetStats(0);
+    sel->GetXaxis()->SetBinLabel(1,"Primary");
+    sel->GetXaxis()->SetBinLabel(2,"Primary Relative");
+    sel->GetXaxis()->SetBinLabel(3,"Hadron");
+    sel->GetXaxis()->SetBinLabel(4,"Low-E e^{#pm}");
+    sel->GetXaxis()->SetBinLabel(5,"Other");
+    sel->GetYaxis()->SetBinLabel(1,"Rad");
+    sel->GetYaxis()->SetBinLabel(2,"EDep");
+    sel->GetYaxis()->SetBinLabel(3,"Time");
+    sel->GetYaxis()->SetBinLabel(4,"Bkg");
+    sel->GetYaxis()->SetBinLabel(5,"BkgClust");
+    sel->GetYaxis()->SetBinLabel(6,"All");
+    selfrac->GetXaxis()->SetBinLabel(1,"Primary");
+    selfrac->GetXaxis()->SetBinLabel(2,"Primary Relative");
+    selfrac->GetXaxis()->SetBinLabel(3,"Hadron");
+    selfrac->GetXaxis()->SetBinLabel(4,"Low-E e^{#pm}");
+    selfrac->GetXaxis()->SetBinLabel(5,"Other");
+    selfrac->GetYaxis()->SetBinLabel(1,"Rad");
+    selfrac->GetYaxis()->SetBinLabel(2,"EDep");
+    selfrac->GetYaxis()->SetBinLabel(3,"Time");
+    selfrac->GetYaxis()->SetBinLabel(4,"Bkg");
+    selfrac->GetYaxis()->SetBinLabel(5,"BkgClust");
+    selfrac->GetYaxis()->SetBinLabel(6,"All");
+    std::vector<TCut> select = {rsel,esel,tsel,bkg,bkgc,rsel+esel+tsel+bkg};
+    std::vector<TCut> mcpart = {pri,prirel,Had,lowEe,other};
+    for(int isel=0;isel<nsel;++isel){
+      for(int ipart=0;ipart<npart;++ipart){
+        char value[10];
+        snprintf(value,10,"%i:%i",isel,ipart);
+//        std::cout << "value " << value << std::endl;
+//        std::cout << "cut" << select[isel]+mcpart[ipart] << std::endl;
+        // using bkg==0 doesn't make sense for lowEe, flipping the sign
+        if(isel == 3 && ipart == 3) CHD->Project("+sel",value,"nsh"*("bkg"+mcpart[ipart]));
+        else CHD->Project("+sel",value,"nsh"*(select[isel]+mcpart[ipart]));
+        CHD->Project("+norm",value,"nsh"*(mcpart[ipart]));
+      }
+    }
+    selfrac->Divide(sel,norm);
+    gStyle->SetPaintTextFormat("4.4f");
+    TCanvas* fcan = new TCanvas("fcan","fcan",1200,600);
+    fcan->Divide(2,1);
+    fcan->cd(1);
+    sel->Draw("box");
+    fcan->cd(2);
+    selfrac->Draw("box");
+    selfrac->Draw("textsame");
+  }
+}
